// src/api/model/asociacionesResponse.ts
/**
 * Contador Playbacks API
 * API para la app contador-playbacks
 *
 * OpenAPI spec version: 1.0.0
 * Contact: transformaciondigital@hogueras.es
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Asociacion } from './asociacion';
import { Response200 } from './response200';

export interface AsociacionesResponse { 
    status?: Response200;
    participants?: Array<Asociacion>;
}